backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg[binary]==3.1.19
python-dotenv==1.0.1
django-environ==0.11.2
boto3==1.34.131
django-storages==1.14.3
gunicorn==22.0.0
